getData

This function allows R users to retrieve match statistics and betting odds for English Premier League games over the past 17 seasons. The data is retrieved from DataHub.io and football-data.co.uk

getData(season, bet = TRUE, team = "all")

Arguments

season
The season argument allows the user to specify for which season he or she wants to get EPL data for. Ensure that season is in the correct character string format, e.g. "1819" for 2018 - 2019 Season; and "1718" for 2017 - 2018 season.

bet
The bet argument allows the user to specify whether he or she wants to retrieve betting data. The default setting for bet is TRUE. If the user only wants the match statistics, set bet to FALSE.

team
The team argument allows the user to filter the data to find match statistics and betting data that he or she is only interested in. The default stting is "all", which retrieves data for all teams in the English Premier league for the specified season. If the user wants to look at only one specific team he or she can use this argument. Ensure that team name is in correct character string format, e.g. "Chelsea" or "Man City".

Value

A dataframe with match statistics and / or betting odds for the specified Season.

Details

The function allows the user to input three arguments. The first argument is season, which specifies the EPL Season that the user wants to retrieve data for. The second argument is bet, which specifies whether the user wants to retrieve additional betting odds data for the matches. The third argument is team, which specifies whether the user wants to retrieve data only for a particular team.

Examples

getData(season = "1819", bet = FALSE, team = "Arsenal")

#>          Date       HomeTeam       AwayTeam FTHG FTAG FTR HTHG HTAG HTR
#> 1  12/08/2018        Arsenal       Man City    0    2   A    0    1   A
#> 2  18/08/2018        Chelsea        Arsenal    3    2   H    2    2   D
#> 3  25/08/2018        Arsenal       West Ham    3    1   H    1    1   D
#> 4  02/09/2018        Cardiff        Arsenal    2    3   A    1    1   D
#> 5  15/09/2018      Newcastle        Arsenal    1    2   A    0    0   D
#> 6  23/09/2018        Arsenal        Everton    2    0   H    0    0   D
#> 7  29/09/2018        Arsenal        Watford    2    0   H    0    0   D
#> 8  07/10/2018         Fulham        Arsenal    1    5   A    1    1   D
#> 9  22/10/2018        Arsenal      Leicester    3    1   H    1    1   D
#> 10 28/10/2018 Crystal Palace        Arsenal    2    2   D    1    0   H
#> 11 03/11/2018        Arsenal      Liverpool    1    1   D    0    0   D
#> 12 11/11/2018        Arsenal         Wolves    1    1   D    0    1   A
#> 13 25/11/2018    Bournemouth        Arsenal    1    2   A    1    1   D
#> 14 02/12/2018        Arsenal      Tottenham    4    2   H    1    2   A
#> 15 05/12/2018     Man United        Arsenal    2    2   D    1    1   D
#> 16 08/12/2018        Arsenal   Huddersfield    1    0   H    0    0   D
#> 17 16/12/2018    Southampton        Arsenal    3    2   H    2    1   H
#> 18 22/12/2018        Arsenal        Burnley    3    1   H    1    0   H
#> 19 26/12/2018       Brighton        Arsenal    1    1   D    1    1   D
#> 20 29/12/2018      Liverpool        Arsenal    5    1   H    4    1   H
#> 21 01/01/2019        Arsenal         Fulham    4    1   H    1    0   H
#> 22 12/01/2019       West Ham        Arsenal    1    0   H    0    0   D
#> 23 19/01/2019        Arsenal        Chelsea    2    0   H    2    0   H
#> 24 29/01/2019        Arsenal        Cardiff    2    1   H    0    0   D
#> 25 03/02/2019       Man City        Arsenal    3    1   H    2    1   H
#> 26 09/02/2019   Huddersfield        Arsenal    1    2   A    0    2   A
#> 27 24/02/2019        Arsenal    Southampton    2    0   H    2    0   H
#> 28 27/02/2019        Arsenal    Bournemouth    5    1   H    2    1   H
#> 29 02/03/2019      Tottenham        Arsenal    1    1   D    0    1   A
#> 30 10/03/2019        Arsenal     Man United    2    0   H    1    0   H
#> 31 01/04/2019        Arsenal      Newcastle    2    0   H    1    0   H
#> 32 07/04/2019        Everton        Arsenal    1    0   H    1    0   H
#> 33 15/04/2019        Watford        Arsenal    0    1   A    0    1   A
#> 34 21/04/2019        Arsenal Crystal Palace    2    3   A    0    1   A
#> 35 24/04/2019         Wolves        Arsenal    3    1   H    3    0   H
#> 36 28/04/2019      Leicester        Arsenal    3    0   H    0    0   D
#> 37 05/05/2019        Arsenal       Brighton    1    1   D    1    0   H
#> 38 12/05/2019        Burnley        Arsenal    1    3   A    0    0   D
#>       Referee HS AS HST AST HF AF HC AC HY AY HR AR
#> 1    M Oliver  9 17   3   8 11 14  2  9  2  2  0  0
#> 2  M Atkinson 24 15  11   6 12  9  5  1  0  2  0  0
#> 3     G Scott 17 13  10   5 16 13 10  2  1  3  0  0
#> 4    A Taylor 14 17   3  11 12 14  3  9  3  4  0  0
#> 5   L Probert  4 12   2   2 13 11 10  4  0  0  0  0
#> 6      J Moss  9  9   5   6 17 12  5  9  2  1  0  0
#> 7    A Taylor  9 13   2   4 11 17  6  6  2  2  0  0
#> 8   P Tierney 21  9   4   7 11 12  4  2  2  0  0  0
#> 9  C Kavanagh 19  8   6   2 10 10  6  4  2  2  0  0
#> 10 M Atkinson 16  7   3   2 10 16  6  4  1  2  0  0
#> 11 A Marriner 12 13   4   4  7  7  5  8  1  1  0  0
#> 12  S Attwell 10 12   3   5  9 16 11  2  2  2  0  0
#> 13   C Pawson 11 20   5   4  6  9  5  8  2  1  0  0
#> 14     M Dean 22 11   7   6 15 17  8  5  3  3  0  1
#> 15 A Marriner 10  9   7   4 13 10  4  4  3  3  0  0
#> 16  P Tierney 14  6   2   0 13 20  7  1  5  4  0  0
#> 17 C Kavanagh 12 13   7   4 12 10  4  5  3  1  0  0
#> 18   K Friend 10  7   6   2 10 14  1  3  2  5  0  0
#> 19   A Taylor 12  7   3   4 10  4  4  9  2  1  0  0
#> 20   M Oliver 15  8  10   2  8 13  5  3  1  2  0  0
#> 21    G Scott 16  9   9   4  7 12  8  3  0  1  0  0
#> 22     J Moss 11 11   3   2  7 10  7  3  0  2  0  0
#> 23   A Taylor 13 13   5   1 13 15  5  6  0  2  0  0
#> 24     M Dean 15 19   4   2 14 12  4  7  3  3  0  0
#> 25 M Atkinson 19  4  12   2 11  8  4  2  1  1  0  0
#> 26     J Moss 15  9   6   4 17 12  5  0  3  2  0  0
#> 27    G Scott 12 10   4   4  7 14  6  4  0  0  0  0
#> 28 C Kavanagh 16 13   6   5 11  9  9  4  2  2  0  0
#> 29   A Taylor 10  9   3   4 15 14  3  4  3  2  0  1
#> 30     J Moss 14 14   3   4 12 18  5  2  2  2  0  0
#> 31   A Taylor  7  3   4   1 11 10  6  2  2  0  0  0
#> 32   K Friend 23  7   6   2  8  9  9  6  1  4  0  0
#> 33   C Pawson 11 19   3   6 12  8  6  4  2  0  1  0
#> 34     J Moss 12 16   5   7 15 12  8  3  4  1  0  0
#> 35  S Attwell 11 11   3   1 12  9  5  5  2  3  0  0
#> 36   M Oliver 24  6  12   1 13 13  8  6  3  1  0  1
#> 37   A Taylor 20 11   8   5  9 14 16  3  5  2  0  0
#> 38     M Dean 14 17   5   6 11  3  4  5  5  1  0  0

getData(season = "1617", bet = TRUE, team = "Chelsea")

#>        Date       HomeTeam       AwayTeam FTHG FTAG FTR HTHG HTAG HTR
#> 1  15/08/16        Chelsea       West Ham    2    1   H    0    0   D
#> 2  20/08/16        Watford        Chelsea    1    2   A    0    0   D
#> 3  27/08/16        Chelsea        Burnley    3    0   H    2    0   H
#> 4  11/09/16        Swansea        Chelsea    2    2   D    0    1   A
#> 5  16/09/16        Chelsea      Liverpool    1    2   A    0    2   A
#> 6  24/09/16        Arsenal        Chelsea    3    0   H    3    0   H
#> 7  01/10/16           Hull        Chelsea    0    2   A    0    0   D
#> 8  15/10/16        Chelsea      Leicester    3    0   H    2    0   H
#> 9  23/10/16        Chelsea     Man United    4    0   H    2    0   H
#> 10 30/10/16    Southampton        Chelsea    0    2   A    0    1   A
#> 11 05/11/16        Chelsea        Everton    5    0   H    3    0   H
#> 12 20/11/16  Middlesbrough        Chelsea    0    1   A    0    1   A
#> 13 26/11/16        Chelsea      Tottenham    2    1   H    1    1   D
#> 14 03/12/16       Man City        Chelsea    1    3   A    1    0   H
#> 15 11/12/16        Chelsea      West Brom    1    0   H    0    0   D
#> 16 14/12/16     Sunderland        Chelsea    0    1   A    0    1   A
#> 17 17/12/16 Crystal Palace        Chelsea    0    1   A    0    1   A
#> 18 26/12/16        Chelsea    Bournemouth    3    0   H    1    0   H
#> 19 31/12/16        Chelsea          Stoke    4    2   H    1    0   H
#> 20 04/01/17      Tottenham        Chelsea    2    0   H    1    0   H
#> 21 14/01/17      Leicester        Chelsea    0    3   A    0    1   A
#> 22 22/01/17        Chelsea           Hull    2    0   H    1    0   H
#> 23 31/01/17      Liverpool        Chelsea    1    1   D    0    1   A
#> 24 04/02/17        Chelsea        Arsenal    3    1   H    1    0   H
#> 25 12/02/17        Burnley        Chelsea    1    1   D    1    1   D
#> 26 25/02/17        Chelsea        Swansea    3    1   H    1    1   D
#> 27 06/03/17       West Ham        Chelsea    1    2   A    0    1   A
#> 28 18/03/17          Stoke        Chelsea    1    2   A    1    1   D
#> 29 01/04/17        Chelsea Crystal Palace    1    2   A    1    2   A
#> 30 05/04/17        Chelsea       Man City    2    1   H    2    1   H
#> 31 08/04/17    Bournemouth        Chelsea    1    3   A    1    2   A
#> 32 16/04/17     Man United        Chelsea    2    0   H    1    0   H
#> 33 25/04/17        Chelsea    Southampton    4    2   H    2    1   H
#> 34 30/04/17        Everton        Chelsea    0    3   A    0    0   D
#> 35 08/05/17        Chelsea  Middlesbrough    3    0   H    2    0   H
#> 36 12/05/17      West Brom        Chelsea    0    1   A    0    0   D
#> 37 15/05/17        Chelsea        Watford    4    3   H    2    1   H
#> 38 21/05/17        Chelsea     Sunderland    5    1   H    1    1   D
#>          Referee HS AS HST AST HF AF HC AC HY AY HR AR B365H B365D B365A  IWH
#> 1       A Taylor 16  7   6   3 16 16  7  1  5  2  0  0  1.65  4.00  6.00 1.65
#> 2         J Moss  6 13   2   4 20 10  0  5  4  3  0  0  5.50  3.80  1.73 4.20
#> 3  M Clattenburg 22  6  10   0  6  9  9  5  2  2  0  0  1.25  5.50 15.00 1.25
#> 4     A Marriner  6 28   2   7 17  8  1 11  3  4  0  0  5.75  4.00  1.67 5.10
#> 5     M Atkinson 12 13   4   5  6 13  6  4  1  1  0  0  2.25  3.50  3.40 2.30
#> 6       M Oliver 14  9   5   2  9 11  9  1  0  2  0  0  2.40  3.50  3.10 2.35
#> 7       A Taylor  8 22   3   9 13 15  5  7  2  2  0  0  7.50  4.50  1.50 6.50
#> 8     A Marriner 16  5   6   0  6 12  8  5  1  1  0  0  1.57  4.50  6.00 1.60
#> 9     M Atkinson 14 16   6   5  7 17  5  7  3  2  0  0  2.15  3.40  3.80 2.20
#> 10       M Jones  9 13   1   7 13 16  7  3  0  0  0  0  3.20  3.40  2.40 3.30
#> 11      R Madley 21  1   9   0  5 16  4  2  0  3  0  0  1.53  4.33  7.00 1.55
#> 12        J Moss 12 13   1   3 13 10  2  3  2  3  0  0  7.00  4.00  1.57 6.50
#> 13      M Oliver  9 12   5   6 13 16  3  8  2  1  0  0  1.75  3.80  5.25 1.75
#> 14      A Taylor 14 10   5   4 14  9  9  2  2  3  2  0  2.15  3.60  3.60 2.00
#> 15        M Dean 12  6   2   1  6 10  3  1  2  4  0  0  1.25  6.25 15.00 1.30
#> 16   N Swarbrick  8 19   3   6 15 10  5  5  3  2  0  0 10.00  5.75  1.33 9.50
#> 17        J Moss  6 13   2   6 12 15  5  3  2  3  0  0  6.50  4.20  1.57 6.50
#> 18       M Jones 14  7   4   3 13 14  5  5  1  1  0  0  1.36  5.00  8.50 1.35
#> 19      R Madley 18  5   7   2 12 10  8  3  3  2  0  0  1.25  6.50 14.00 1.27
#> 20    M Atkinson  9 11   2   2  9  8  1  6  3  2  0  0  2.63  3.25  2.63 3.15
#> 21    A Marriner  7  8   2   3  8  9  3  5  1  0  0  0  6.00  4.10  1.62 5.00
#> 22   N Swarbrick  9  9   5   4 12 10  9  7  1  3  0  0  1.18  8.00 19.00 1.15
#> 23 M Clattenburg  7  8   3   2 13  8  3  3  2  1  0  0  2.65  3.25  2.90 2.60
#> 24    M Atkinson 13  9   6   5 14  8 10 13  1  1  0  0  1.95  3.60  4.20 2.10
#> 25      K Friend  6 13   4   2 12 11  1  1  3  2  0  0  9.00  5.00  1.40 7.50
#> 26   N Swarbrick 16  3   5   2 11 14  8  1  1  3  0  0  1.25  6.50 15.00 1.22
#> 27    A Marriner 11  9   2   4  8  8  6  4  0  1  0  0  6.00  4.10  1.62 5.70
#> 28      A Taylor  5 20   1   7 16 11  5 11  4  2  1  0  6.50  4.00  1.62 5.30
#> 29      C Pawson 24  8  11   3 12 14  9  3  3  2  0  0  1.30  5.75 12.00 1.25
#> 30        M Dean 10 17   4   7 10 14  2  9  1  3  0  0  2.40  3.50  3.10 2.20
#> 31    A Marriner 12 15   1   5  7 11  4  4  2  3  0  0  6.50  4.00  1.60 6.20
#> 32      R Madley  9  5   3   0 16 20  1  3  3  3  0  0  2.55  3.20  2.80 2.65
#> 33       L Mason 17 12   7   4 13  9  6  7  2  2  0  0  1.45  4.60  8.50 1.45
#> 34        J Moss 12 11   4   5 13 11  2  4  3  4  0  0  3.50  3.60  2.00 3.80
#> 35      C Pawson 21  2   7   1  8 21  8  1  0  2  0  0  1.17  9.00 19.00 1.15
#> 36      M Oliver  7 24   2   5 13  8  5  8  3  0  0  0 10.00  5.25  1.36 9.50
#> 37       L Mason 24  9   9   3  7 12  8  3  2  4  0  1  1.25  6.75 13.00 1.33
#> 38   N Swarbrick 28  7   8   3  8 15 11  1  1  1  0  0  1.14  9.00 23.00 1.20
#>     IWD   IWA   WHH  WHD   WHA
#> 1  3.60  5.10  1.67 3.60  5.50
#> 2  3.50  1.80  5.50 3.30  1.73
#> 3  5.50 10.30  1.29 5.00 12.00
#> 4  3.60  1.65  6.00 3.60  1.62
#> 5  3.30  3.15  2.30 3.40  3.40
#> 6  3.30  3.05  2.40 3.20  3.20
#> 7  4.20  1.50  7.00 4.33  1.50
#> 8  3.90  5.70  1.60 3.80  6.50
#> 9  3.30  3.35  2.25 3.10  3.60
#> 10 3.35  2.20  3.20 3.20  2.40
#> 11 4.00  6.20  1.57 3.75  7.00
#> 12 4.20  1.50  6.50 4.00  1.57
#> 13 3.60  4.80  1.80 3.50  5.00
#> 14 3.40  3.80  2.20 3.30  3.50
#> 15 5.50  9.50  1.29 5.50 13.00
#> 16 5.50  1.30  9.00 5.00  1.33
#> 17 4.20  1.50  6.50 3.75  1.62
#> 18 5.30  8.00  1.40 4.80  8.50
#> 19 5.80 10.50  1.29 5.50 13.00
#> 20 3.30  2.30  2.90 3.10  2.62
#> 21 4.00  1.65  6.00 3.60  1.67
#> 22 7.50 18.00  1.20 7.00 15.00
#> 23 3.20  2.80  2.60 3.20  3.00
#> 24 3.40  3.50  2.00 3.30  4.20
#> 25 4.80  1.40  8.00 4.50  1.44
#> 26 6.50 12.00  1.25 6.00 13.00
#> 27 3.90  1.60  6.00 4.00  1.62
#> 28 3.80  1.65  6.00 4.00  1.62
#> 29 6.00 11.00  1.29 5.50 12.00
#> 30 3.35  3.30  2.40 3.30  3.10
#> 31 4.00  1.55  6.00 3.60  1.62
#> 32 3.30  2.65  2.50 3.20  2.88
#> 33 4.40  7.20  1.44 4.33  8.00
#> 34 3.40  2.00  3.50 3.50  2.05
#> 35 7.50 18.00  1.14 8.50 19.00
#> 36 5.50  1.30 10.00 4.40  1.36
#> 37 5.50  8.00  1.25 6.00 11.00
#> 38 6.50 15.00  1.15 7.50 17.00

getData(season = "1415", team = "Man City")

#>        Date       HomeTeam       AwayTeam FTHG FTAG FTR HTHG HTAG HTR
#> 1  17/08/14      Newcastle       Man City    0    2   A    0    1   A
#> 2  25/08/14       Man City      Liverpool    3    1   H    1    0   H
#> 3  30/08/14       Man City          Stoke    0    1   A    0    0   D
#> 4  13/09/14        Arsenal       Man City    2    2   D    0    1   A
#> 5  21/09/14       Man City        Chelsea    1    1   D    0    0   D
#> 6  27/09/14           Hull       Man City    2    4   A    2    2   D
#> 7  04/10/14    Aston Villa       Man City    0    2   A    0    0   D
#> 8  18/10/14       Man City      Tottenham    4    1   H    2    1   H
#> 9  25/10/14       West Ham       Man City    2    1   H    1    0   H
#> 10 02/11/14       Man City     Man United    1    0   H    0    0   D
#> 11 08/11/14            QPR       Man City    2    2   D    1    1   D
#> 12 22/11/14       Man City        Swansea    2    1   H    1    1   D
#> 13 30/11/14    Southampton       Man City    0    3   A    0    0   D
#> 14 03/12/14     Sunderland       Man City    1    4   A    1    2   A
#> 15 06/12/14       Man City        Everton    1    0   H    1    0   H
#> 16 13/12/14      Leicester       Man City    0    1   A    0    1   A
#> 17 20/12/14       Man City Crystal Palace    3    0   H    0    0   D
#> 18 26/12/14      West Brom       Man City    1    3   A    0    3   A
#> 19 28/12/14       Man City        Burnley    2    2   D    2    0   H
#> 20 01/01/15       Man City     Sunderland    3    2   H    0    0   D
#> 21 10/01/15        Everton       Man City    1    1   D    0    0   D
#> 22 18/01/15       Man City        Arsenal    0    2   A    0    1   A
#> 23 31/01/15        Chelsea       Man City    1    1   D    1    1   D
#> 24 07/02/15       Man City           Hull    1    1   D    0    1   A
#> 25 11/02/15          Stoke       Man City    1    4   A    1    1   D
#> 26 21/02/15       Man City      Newcastle    5    0   H    3    0   H
#> 27 01/03/15      Liverpool       Man City    2    1   H    1    1   D
#> 28 04/03/15       Man City      Leicester    2    0   H    1    0   H
#> 29 14/03/15        Burnley       Man City    1    0   H    0    0   D
#> 30 21/03/15       Man City      West Brom    3    0   H    2    0   H
#> 31 06/04/15 Crystal Palace       Man City    2    1   H    1    0   H
#> 32 12/04/15     Man United       Man City    4    2   H    2    1   H
#> 33 19/04/15       Man City       West Ham    2    0   H    2    0   H
#> 34 25/04/15       Man City    Aston Villa    3    2   H    1    0   H
#> 35 03/05/15      Tottenham       Man City    0    1   A    0    1   A
#> 36 10/05/15       Man City            QPR    6    0   H    2    0   H
#> 37 17/05/15        Swansea       Man City    2    4   A    1    2   A
#> 38 24/05/15       Man City    Southampton    2    0   H    1    0   H
#>          Referee HS AS HST AST HF AF HC AC HY AY HR AR B365H B365D B365A  IWH
#> 1     M Atkinson 12 13   0   5  8 11  3  3  1  5  0  0  5.25  4.33  1.67 4.90
#> 2       M Oliver  9 11   4   3 13  7  6  7  1  1  0  0  1.85  4.00  4.33 1.85
#> 3        L Mason 16  7   2   2 14  5 11  4  2  1  0  0  1.18  8.00 18.00 1.20
#> 4  M Clattenburg 15 15   6   8 11 15  0  3  3  4  0  0  3.00  3.40  2.50 2.75
#> 5         M Dean 16  6   4   2 16 12 14  2  3  4  1  0  2.38  3.50  3.10 2.40
#> 6       A Taylor 10 20   5   6 10  6  3  8  0  2  0  0  6.50  4.20  1.57 5.50
#> 7          C Foy  6 27   1   7  4  6  0  7  0  1  0  0  7.50  4.50  1.50 5.60
#> 8         J Moss 20 19  12   7 11  9  3  4  1  2  0  1  1.45  4.75  8.00 1.45
#> 9     M Atkinson 12 21   4   5 12 14  4  8  2  1  0  0  6.50  4.50  1.55 5.60
#> 10      M Oliver 16  9   6   2 15  9  7  4  3  1  0  1  1.91  3.75  4.33 1.90
#> 11        M Dean 17 23   4   6  9 10  7  7  4  2  0  0  7.00  4.75  1.50 7.00
#> 12   N Swarbrick 23  8  10   3 13  9  7  3  2  1  0  0  1.40  5.25  8.50 1.35
#> 13       M Jones 13 15   4   6 10 13  8  6  1  2  0  1  3.20  3.60  2.30 3.10
#> 14      C Pawson  8 15   2   4  9  8  2  0  1  1  0  0  7.50  4.50  1.50 6.10
#> 15    A Marriner 19  9   3   2  8 11 11  7  3  3  0  0  1.50  4.75  7.00 1.57
#> 16        J Moss  7  9   1   4 15 13  8  2  0  2  0  0  7.00  4.75  1.50 7.00
#> 17        P Dowd 15  6   3   1 17  5  8  6  2  1  0  0  1.29  6.00 12.00 1.30
#> 18 M Clattenburg 18  9   6   5  5  9 10  6  0  1  0  0  6.50  4.50  1.53 7.00
#> 19      K Friend 17 12   6   3 18  9 10  3  4  1  0  0  1.18  8.00 19.00 1.17
#> 20        R East 32  4  12   4 14  7 10  4  2  3  0  0  1.22  7.00 15.00 1.25
#> 21    M Atkinson 11 18   4   2  9 10  4  9  1  2  0  0  4.50  3.75  1.85 3.80
#> 22        M Dean 12  9   4   3 12  5 16  3  3  3  0  0  1.80  4.00  4.50 1.80
#> 23 M Clattenburg  3 10   2   5 12 14  1  4  0  2  0  0  2.15  3.50  3.60 1.90
#> 24        J Moss 18  7   5   3  9  8 13  1  2  4  0  0  1.18  8.00 19.00 1.20
#> 25       L Mason 17 14   5   8  6 11  3  2  1  2  0  0  5.75  3.80  1.70 4.70
#> 26         C Foy 16  8   9   1  9  8  7  3  3  2  0  0  1.30  6.00 11.00 1.30
#> 27 M Clattenburg 11  8   5   1  8 16  3  0  1  3  0  0  3.10  3.60  2.38 2.75
#> 28      R Madley 22  6   9   2 11  3 10  3  0  1  0  0  1.25  7.00 13.00 1.25
#> 29    A Marriner 10 21   4   5  6  9  2  6  2  1  0  0  7.50  5.00  1.44 7.00
#> 30   N Swarbrick 43  3  16   0  6  5  9  2  0  0  0  1  1.30  5.50  8.50 1.30
#> 31      M Oliver  5 22   3   4 11 11  4 13  2  1  0  0  5.00  4.00  1.67 5.00
#> 32 M Clattenburg 11 10   7   3  9 16  1  5  0  3  0  0  2.63  3.40  2.88 2.40
#> 33      A Taylor 18  9   3   3 10  6  6  3  2  2  0  0  1.25  6.50 11.00 1.27
#> 34        M Dean 11 13   4   4 13 11  7  6  1  2  0  0  1.29  6.00 12.00 1.27
#> 35    A Marriner 21 15   5   4  6 11  8  7  3  4  0  0  3.75  3.75  2.05 3.60
#> 36        M Dean 23  9  11   3 12  7  7  4  0  1  0  0  1.22  7.50 13.00 1.20
#> 37 M Clattenburg 13 22   9  11  8  8 11  7  0  3  0  0  4.75  4.20  1.75 4.50
#> 38         C Foy 15 13   6   4 13  8  8  4  1  1  0  0  1.57  4.50  6.00 1.50
#>     IWD   IWA  WHH  WHD   WHA
#> 1  3.70  1.65 5.00 3.75  1.67
#> 2  3.45  4.00 1.85 3.40  4.33
#> 3  5.80 13.00 1.20 5.50 21.00
#> 4  3.30  2.40 3.00 3.00  2.50
#> 5  3.30  2.75 2.40 3.10  3.10
#> 6  3.80  1.57 6.50 3.60  1.57
#> 7  3.90  1.55 7.00 3.75  1.53
#> 8  4.00  7.00 1.50 4.00  7.00
#> 9  3.90  1.55 5.50 4.00  1.55
#> 10 3.30  4.00 1.91 3.30  4.20
#> 11 4.00  1.45 7.00 4.00  1.50
#> 12 4.80  7.60 1.40 4.33  9.00
#> 13 3.30  2.20 3.20 3.00  2.38
#> 14 4.00  1.50 6.50 3.80  1.53
#> 15 3.80  5.50 1.53 4.00  6.00
#> 16 4.00  1.45 6.00 4.33  1.50
#> 17 5.00  9.00 1.30 5.00 11.00
#> 18 4.00  1.45 7.00 4.00  1.50
#> 19 6.50 14.00 1.17 6.50 19.00
#> 20 5.50 10.30 1.25 5.00 15.00
#> 21 3.45  1.90 4.33 3.30  1.91
#> 22 3.50  4.20 1.83 3.80  4.00
#> 23 3.45  3.80 2.15 3.20  3.60
#> 24 5.80 13.00 1.20 6.00 17.00
#> 25 3.60  1.70 5.50 3.50  1.70
#> 26 5.00  9.00 1.30 4.50 13.00
#> 27 3.30  2.40 3.10 3.10  2.40
#> 28 5.50 10.30 1.25 5.50 13.00
#> 29 4.00  1.45 7.00 4.50  1.44
#> 30 5.00  9.00 1.33 5.00  9.00
#> 31 3.90  1.60 4.80 3.80  1.70
#> 32 3.30  2.75 2.70 3.10  2.70
#> 33 5.20 10.00 1.29 5.00 12.00
#> 34 5.20 10.00 1.29 4.80 12.00
#> 35 3.30  2.00 3.75 3.30  2.05
#> 36 5.80 13.00 1.22 6.00 13.00
#> 37 3.50  1.75 5.00 3.50  1.75
#> 38 4.00  6.10 1.57 3.80  6.00