teamData

This function allows R users to get season statistics for a specified team and season in the English Premier League. The data is retrieved from DataHub.io and football-data.co.uk. It covers from 2009 to 2019.

teamData(season, team)

Arguments

season
The season argument allows the user to specify for which season he or she wants to get EPL data. Ensure that season is in the correct character string format, e.g. "1819" for 2018 - 2019 Season and "1718" for 2017 - 2018 season.

team
The team argument requires the user to specify the team he or she wants the data for. The first letter of the team name should be capitalized or if the team is name is 2 words, e.g. Man United - you should capitalize the first letter of both words.

Value

A dataframe with summary season statistics for the specified team and season.

Details

The function requires users to input two different arguments. The first argument specifies the EPL Season that he or she wants to retrieve data for. The second argument specifies which team he or she wants to get data for.

Examples

teamData(season = "1819", team = "Liverpool")

#>        Team Season Games Win Draw Loss
#> 1 Liverpool   1819    38  30    7    1

teamData("1617", "Man United")

#>         Team Season Games Win Draw Loss
#> 1 Man United   1617    38  18   15    5